Turn processes of moving and game checking decoupled from house creation
